+++ db/migrate/20200624150605_add_certificate_order_id_to_ahoy_email.rb
@@ -1,0 +1,9 @@
+class AddCertificateOrderIdToAhoyEmail < ActiveRecord::Migration
+  def up
+     add_reference :ahoy_messages, :certificate_order, foreign_key: true
+   end
+
+   def down
+     remove_reference :ahoy_messages, :certificate_order
+   end
+end