+++ CHANGELOG.md
@@ -3,9 +3,19 @@ All notable changes to this project will be documented in this file.
 
 ### PENDING RELEASE (TBD)
 
+### v1.3.1 (June 29, 2020)
+
+- Allow document upload on verified orders
+
+### v1.3.0 (June 29, 2020)
+
 - Created changelog for tracking codebase changes
 - Added test coverage upload for CodeClimate
 - Fixed test coverage upload for CodeClimate
+- Update CertificateOrder.search_with_csr to include contacts
+- Move CertificateOrder scopes to a Concern
+- Lock Rubocop Version to match Code CodeClimate
+- Create a Dashboard for Phone Approval and Verification Management
 
 ### v1.2.0 (June 17, 2020)